+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Changed
+
+- Admin: change the product files form part icon to a file icon
+
 ### Added
 
 - Core: add provider to retrieve subscription options for a given product