+++ CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+
+
 ### Added
 
+- Core: add verbose_name to shop product so we can translate it
 - Core: add middleware provides to the basket command handler
 - Core: add provides to retrieve properties from order or order source
 - Front: render subscription options in basket template
@@ -20,6 +23,7 @@ List all changes after the last release here (newer on top). Each change on a se
 - Front: render basket and order line properties using the `front_line_properties_descriptor` provides
 - Admin: change the product files form part icon to a file icon
 - Core: cache language utils methods using LRU
+
 
 ## [2.3.10] - 2021-01-22