+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
-
+## [2.3.11] - 2021-01-25
 
 ### Added
 
@@ -23,7 +23,6 @@ List all changes after the last release here (newer on top). Each change on a se
 - Front: render basket and order line properties using the `front_line_properties_descriptor` provides
 - Admin: change the product files form part icon to a file icon
 - Core: cache language utils methods using LRU
-
 
 ## [2.3.10] - 2021-01-22